+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Upgrading to 3.8.x versions requires that you upgrade to latest 3.5.x version first.
 
+- Add `MAIL_PROCESS_BEFORE` event to skip processing certain emails, #820
+
 [3.8.11]: https://github.com/eventum/eventum/compare/v3.8.10...master
 
 ## [3.8.10] - 2020-04-29